Ordering added.

+++ src/user/models.py
@@ -18,7 +18,7 @@ class User(models.Model):
     phone_number    = models.CharField(max_length = 20, default = '', blank = True, null = True, unique = False, verbose_name = 'Контактный номер')
     account_id      = models.CharField(max_length = 64, blank = False, null = False, unique = True, verbose_name = 'Аккаунт')
     group_number    = models.CharField(max_length = 4, blank = False, null = False, verbose_name = 'Номер группы')
-    account         = models.PositiveSmallIntegerField(default = 0, blank = False, null = False, verbose_name = 'Счёт')
+    account         = models.DecimalField(default = 0, max_digits = 7, decimal_places = 2, blank = False, null = False, verbose_name = 'Счёт')
     avatar_url      = models.URLField(null = True, blank = True)
     # Two 'blank' (unrequired) values can`t be unique
     email           = models.CharField(default = '', max_length = 64, blank = True, null = False, unique = False, verbose_name = 'Почта')
